Convert to plain Node.js instead of Dockerfile
Instead of using a Dockerfile, make this repository into a Node.js
project that can be deployed in OpenShift with the Node.js S2I build
strategy. Remove the Dockerfile as obsolete. Update the README file
accordingly.

Instead of using NGINX, express is used to serve the static content of
the web page. There is an index.js file for this purpose. Another new
addition is a package.json file that contains dependencies and NPM
scripts for development and server start purposes.

Reveal.js is upgraded to version 3.8.0. This broke one use of emphasis
on one slide, so the emphasis was removed.

There are two helper scripts - build.sh and symlink.sh. The build script
is used to copy over presentation material to the reveal.js directory
for serving. The symlink script is used in development to symlink
presentation content to the reveal.js directory.

A .gitignore file is added so that we can ignore the node_modules
directory created by "npm install".